+++ test/audiovideofacade/DefaultAudioVideoFacade.test.ts
@@ -133,6 +133,12 @@ describe('DefaultAudioVideoFacade', () => {
       assert(spy.calledOnceWith());
     });
 
+    it('will call setVideoMaxBandwidthKbps', () => {
+      const spy = sinon.spy(controller, 'setVideoMaxBandwidthKbps');
+      facade.setVideoMaxBandwidthKbps(100);
+      assert(spy.calledOnce);
+    });
+
     it('will call getRTCPeerConnectionStats', async () => {
       const spy = sinon.spy(controller, 'getRTCPeerConnectionStats');
       await facade.getRTCPeerConnectionStats();
@@ -509,28 +515,44 @@ describe('DefaultAudioVideoFacade', () => {
       assert(spy.calledOnceWith());
     });
 
-    it('will call chooseAudioInputDevice', async () => {
-      const spy = sinon.spy(deviceController, 'chooseAudioInputDevice');
+    it('will call startAudioInput', async () => {
+      const spy = sinon.spy(deviceController, 'startAudioInput');
       const arg1 = '';
       try {
-        await facade.chooseAudioInputDevice(arg1);
+        await facade.startAudioInput(arg1);
       } catch (e) {}
       assert(spy.calledOnceWith(arg1));
     });
 
-    it('will call chooseVideoInputDevice', async () => {
-      const spy = sinon.spy(deviceController, 'chooseVideoInputDevice');
+    it('will call stopAudioInput', async () => {
+      const spy = sinon.spy(deviceController, 'stopAudioInput');
+      try {
+        await facade.stopAudioInput();
+      } catch (e) {}
+      assert(spy.calledOnce);
+    });
+
+    it('will call startVideoInput', async () => {
+      const spy = sinon.spy(deviceController, 'startVideoInput');
       const arg1 = '';
       try {
-        await facade.chooseVideoInputDevice(arg1);
+        await facade.startVideoInput(arg1);
       } catch (e) {}
       assert(spy.calledOnceWith(arg1));
 
       const arg2 = new DefaultVideoTransformDevice(null, '', []);
       try {
-        await facade.chooseVideoInputDevice(arg2);
+        await facade.startVideoInput(arg2);
       } catch (e) {}
       assert(spy.calledWith(arg2));
+    });
+
+    it('will call stopVideoInput', async () => {
+      const spy = sinon.spy(deviceController, 'stopVideoInput');
+      try {
+        await facade.stopVideoInput();
+      } catch (e) {}
+      assert(spy.calledOnce);
     });
 
     it('will call chooseVideoInputQuality', () => {
@@ -538,9 +560,8 @@ describe('DefaultAudioVideoFacade', () => {
       const arg1 = 1;
       const arg2 = 2;
       const arg3 = 3;
-      const arg4 = 4;
-      facade.chooseVideoInputQuality(arg1, arg2, arg3, arg4);
-      assert(spy.calledOnceWith(arg1, arg2, arg3, arg4));
+      facade.chooseVideoInputQuality(arg1, arg2, arg3);
+      assert(spy.calledOnceWith(arg1, arg2, arg3));
     });
 
     it('will call getVideoInputQualitySettings', () => {
@@ -550,10 +571,10 @@ describe('DefaultAudioVideoFacade', () => {
     });
 
     it('will call chooseAudioOutputDevice', async () => {
-      const spy = sinon.spy(deviceController, 'chooseAudioOutputDevice');
+      const spy = sinon.spy(deviceController, 'chooseAudioOutput');
       const arg1 = '';
       try {
-        await facade.chooseAudioOutputDevice(arg1);
+        await facade.chooseAudioOutput(arg1);
       } catch (e) {}
       assert(spy.calledOnceWith(arg1));
     });